+++ tcp-streams/System/IO/Streams/TLS.hs
@@ -15,30 +15,25 @@
 -- @
 --
 module System.IO.Streams.TLS
-  ( -- * client
-    connect
-  , withConnection
-    -- * server
+  ( -- * connection type
+    Connection(..)
+    -- * tcp client
+  , connect
+  , connectTLS
+  , tLsToConnection
+    -- * tcp server
   , accept
-    -- * helpers
-  , tlsToStreams
-  , close
-    -- * re-export helpers
   , module Data.TLSSetting
   ) where
 
 import qualified Control.Exception     as E
-import           Control.Monad         (void)
-import           Data.ByteString       (ByteString)
+import           Data.Connection
 import qualified Data.ByteString       as B
 import qualified Data.ByteString.Char8 as BC
-import           Data.ByteString.Lazy  (fromStrict)
 import           Data.TLSSetting
-import           Network.Socket        (HostName, PortNumber, Socket)
 import qualified Network.Socket        as N
 import           Network.TLS           (ClientParams, Context, ServerParams)
 import qualified Network.TLS           as TLS
-import           System.IO.Streams     (InputStream, OutputStream)
 import qualified System.IO.Streams     as Stream
 import qualified System.IO.Streams.TCP as TCP
 
@@ -46,74 +41,50 @@ import qualified System.IO.Streams.TCP as TCP
 -- | Given an existing TLS 'Context' connection, produces an 'InputStream' \/
 -- 'OutputStream' pair.
 --
-tlsToStreams :: Context             -- ^ TLS connection object
-             -> IO (InputStream ByteString, OutputStream ByteString)
-tlsToStreams ctx = do
+tLsToConnection :: (Context, N.SockAddr)    -- ^ TLS connection / socket address pair
+                -> IO Connection
+tLsToConnection (ctx, addr) = do
     is <- Stream.makeInputStream input
-    os <- Stream.makeOutputStream output
-    return (is, os)
+    return (Connection is write (closeTLS ctx) addr)
   where
-    input = ( do
+    input = (do
         s <- TLS.recvData ctx
         return $! if B.null s then Nothing else Just s
         ) `E.catch` (\(_::E.SomeException) -> return Nothing)
-
-    output Nothing  = return ()
-    output (Just s) = TLS.sendData ctx (fromStrict s)
-{-# INLINABLE tlsToStreams #-}
+    write s = TLS.sendData ctx s
 
 -- | Close a TLS 'Context' and its underlying socket.
 --
-close :: Context -> IO ()
-close ctx = (TLS.bye ctx >> TLS.contextClose ctx) -- sometimes socket was closed before 'TLS.bye'
+closeTLS :: Context -> IO ()
+closeTLS ctx = (TLS.bye ctx >> TLS.contextClose ctx) -- sometimes socket was closed before 'TLS.bye'
     `E.catch` (\(_::E.SomeException) -> return ())   -- so we catch the 'Broken pipe' error here
 
 -- | Convenience function for initiating an TLS connection to the given
 -- @('HostName', 'PortNumber')@ combination.
 --
--- Note that sending an end-of-file to the returned 'OutputStream' will not
--- close the underlying TLS connection; to do that, call 'close'
+-- This operation will throw 'TLS.TLSException' on failure.
 --
--- this operation will throw 'TLS.TLSException' on failure.
---
+connectTLS :: ClientParams         -- ^ check "Data.TLSSetting".
+           -> Maybe String         -- ^ Optional certificate subject name, if set to 'Nothing'
+                                   -- then we will try to verify 'HostName' as subject name.
+           -> N.HostName             -- ^ hostname to connect to
+           -> N.PortNumber           -- ^ port number to connect to
+           -> IO (Context, N.SockAddr)
+connectTLS prms subname host port = do
+    let subname' = maybe host id subname
+        prms' = prms { TLS.clientServerIdentification = (subname', BC.pack (show port)) }
+    (sock, addr) <- TCP.connectSocket host port
+    E.bracketOnError (TLS.contextNew sock prms') closeTLS $ \ ctx -> do
+        TLS.handshake ctx
+        return (ctx, addr)
+
 connect :: ClientParams         -- ^ check "Data.TLSSetting".
         -> Maybe String         -- ^ Optional certificate subject name, if set to 'Nothing'
                                 -- then we will try to verify 'HostName' as subject name.
-        -> HostName             -- ^ hostname to connect to
-        -> PortNumber           -- ^ port number to connect to
-        -> IO (InputStream ByteString, OutputStream ByteString, Context)
-connect prms subname host port = do
-    let subname' = maybe host id subname
-        prms' = prms { TLS.clientServerIdentification = (subname', BC.pack (show port)) }
-    sock <- TCP.connectSocket host port
-    E.bracketOnError (TLS.contextNew sock prms') close $ \ ctx -> do
-        TLS.handshake ctx
-        (is, os) <- tlsToStreams ctx
-        return (is, os, ctx)
-
-
--- | Convenience function for initiating an TLS connection to the given
--- @('HostName', 'PortNumber')@ combination. The socket and TLS connection are
--- closed and deleted after the user handler runs.
---
-withConnection :: ClientParams
-               -> Maybe HostName
-               -> HostName
-               -> PortNumber
-               -> (InputStream ByteString -> OutputStream ByteString -> Context -> IO a)
-                       -- ^ Action to run with the new connection
-               -> IO a
-withConnection prms subname host port action =
-    E.bracket (connect prms subname host port) cleanup go
-
-  where
-    go (is, os, ctx) = action is os ctx
-
-    cleanup (_, os, ctx) = E.mask_ $
-        eatException $! Stream.write Nothing os >> close ctx
-
-    eatException m = void m `E.catch` (\(_::E.SomeException) -> return ())
-
+        -> N.HostName             -- ^ hostname to connect to
+        -> N.PortNumber           -- ^ port number to connect to
+        -> IO Connection
+connect prms subname host port = connectTLS prms subname host port >>= tLsToConnection
 
 -- | Accept a new connection from remote client, return a 'InputStream' / 'OutputStream'
 -- pair and remote 'N.SockAddr', you should call 'TCP.bindAndListen' first.
@@ -121,11 +92,12 @@ withConnection prms subname host port action =
 -- this operation will throw 'TLS.TLSException' on failure.
 --
 accept :: ServerParams              -- ^ check "Data.TLSSetting".
-       -> Socket                    -- ^ the listening 'Socket'.
-       -> IO (InputStream ByteString, OutputStream ByteString, Context, N.SockAddr)
+       -> N.Socket                    -- ^ the listening 'Socket'.
+       -> IO Connection
 accept prms sock = do
-    (sock', sockAddr) <- N.accept sock
-    E.bracketOnError (TLS.contextNew sock' prms) close $ \ ctx -> do
+    (sock', addr) <- N.accept sock
+    E.bracketOnError (TLS.contextNew sock' prms) closeTLS $ \ ctx -> do
         TLS.handshake ctx
-        (is, os) <- tlsToStreams ctx
-        return (is, os, ctx, sockAddr)
+        conn <- tLsToConnection (ctx, addr)
+        return conn
+