+++ tcp-streams/ChangeLog.md
@@ -1,5 +1,24 @@
 # Revision history for tcp-streams
 
+## 1.0.0.0
+
+Major rewrite using new Connection interface.
+
+* Introduce `Connection` in `Data.Connection` to leaverage vectorized IO.
+* Now `connect`, `accept` return `Connection`, remove `withXXX` functions.
+* Change exception behavior to work better with `bracket`.
+* `bindAndListen` 's type changed from `PortNumber -> Int -> IO Socket` to `Int -> PortNumber -> IO Socket`.
+* Update built-in mozilla CA list(2017/01/18).
+
+
+## 0.7.0.0
+
+* Add built-in timeout support
+* Change behavior using io-streams' convention: 
+    * reading `InputStream` will throw exception, and the socket won't be closed.
+    * writing to `OutputStream` will not cause a write until flush.
+    * writing empty `ByteString` to `OutputStream` will do a flush.
+
 ## 0.6.0.0
 
 * Update built-in mozilla CA list(2016/11/02).